+++ src/jsMain/kotlin/dev/inmo/jsuikit/modifiers/UIKitCustom.kt
@@ -1,0 +1,6 @@
+package dev.inmo.jsuikit.modifiers
+
+class UIKitCustom(
+    override val classes: Array<String> = emptyArray(),
+    override val otherAttrs: Map<String, String> = emptyMap()
+) : UIKitModifier