+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Reorder arguments in `DefaultComment` fun
 * Add `Vertical` and `Horizontal` members in margins
+* Add `afterHeaderBuilder` and `beforeFooterBuilder` properties in `Dialog` fun
 
 ## 0.0.35