+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Reorder arguments in `DefaultComment` fun
 * Add `Vertical` and `Horizontal` members in margins
 * Add `afterHeaderBuilder` and `beforeFooterBuilder` properties in `Dialog` fun
+* Add `UIKitCustom` to be able for simple creating of custom modifiers
 
 ## 0.0.35