+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add `Vertical` and `Horizontal` members in margins
 * Add `afterHeaderBuilder` and `beforeFooterBuilder` properties in `Dialog` fun
 * Add `UIKitCustom` to be able for simple creating of custom modifiers
+* Add support of `UIKitForm`
 
 ## 0.0.35